requirements.txt
celery==5.0.5
Django==3.2
django-debug-toolbar==3.2
django-environ==0.4.5
django-extensions==3.1.1
django-registration-redux==2.9
django-webpack-loader==0.7.0
flake8==3.9.0
gunicorn==20.1.0
psycopg2-binary==2.8.6
pytest==6.2.2
pytest-django==4.1.0
redis==3.5.3
djangorestframework==3.12.4
markdown==3.3.4
django-filter==2.4.0
Pillow==8.2.0
